Refactor helpers into graphics utilities
